+++ internal/machine/spectrum/bus_test.go
@@ -8,4 +8,5 @@ import (
 
 func TestBus(t *testing.T) {
 	dsl.RunScenarios(t, busScenarios)
+	dsl.RunScenarios(t, keyboardScenarios)
 }